refactor: separate files for entry cards
