+++ lib/features/journal/ui/widgets/list_cards/card_image_widget.dart
