+++ lib/features/journal/ui/widgets/list_cards/card_wrapper_widget.dart
@@ -1,0 +1,31 @@
+import 'package:flutter/material.dart';
+import 'package:lotti/classes/journal_entities.dart';
+import 'package:lotti/features/journal/ui/widgets/list_cards/journal_card.dart';
+import 'package:lotti/features/journal/ui/widgets/list_cards/journal_image_card.dart';
+import 'package:lotti/features/journal/ui/widgets/list_cards/task_list_card_card.dart';
+
+class CardWrapperWidget extends StatelessWidget {
+  const CardWrapperWidget({
+    required this.item,
+    required this.taskAsListView,
+    super.key,
+  });
+
+  final JournalEntity item;
+  final bool taskAsListView;
+
+  @override
+  Widget build(BuildContext context) {
+    return item.maybeMap(
+      journalImage: (JournalImage image) => JournalImageCard(item: image),
+      task: (Task task) {
+        if (taskAsListView) {
+          return TaskListCard(task: task);
+        } else {
+          return JournalCard(item: task);
+        }
+      },
+      orElse: () => JournalCard(item: item),
+    );
+  }
+}
